4.21

+++ osa4/controllers/blogs.js
@@ -41,6 +41,17 @@ blogsRouter.post('/', async (request, response) => {
 
 // Delete a blog
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(401).json({ error: 'unauthorized user' })
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
